Fixed :d command
The :d command now drops as many elements as the number passed as its
argument. If there are not enough elements in the stack, the command
fails without dropping any element from the stack. If a negative number
is passed, the whole stack is dropped. Command :D is removed
accordingly.